pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "nats-py>=2.10.0",
    "pydantic>=2.11.5",
    "qdrant-client[fastembed]>=1.14.2",
    "uvicorn>=0.34.2",
]
[tool.basedpyright]
typeCheckingMode = "standard"
